+++ src/components/product/ProductShelf/ProductShelf.tsx
@@ -1,4 +1,4 @@
-import { useProductsQuery } from 'src/sdk/product/useProductsQuery'
+import productsQuery from 'src/mocks/productsQuery.json'
 import type { ProductsQueryQueryVariables } from '@generated/graphql'
 import Carousel from '@acctglobal/carousel-universal'
 import useWindowDimensions from 'src/sdk/utils/useWindowDimensions'
@@ -23,10 +23,9 @@ function ProductShelf({
   cardsQuantity,
   title,
   withDivisor = false,
-  ...variables
 }: ProductShelfProps) {
   const { isTablet } = useWindowDimensions()
-  const products = useProductsQuery(variables)
+  const { products } = productsQuery.data.search
 
   if (products?.edges.length === 0) {
     return null